novel-editor/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
httpx==0.25.2
pydantic[email]==2.5.0
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
cryptography==42.0.8
websockets==12.0
